style: Resize images to fit better with autolayout
my Mentor Martin suggested to rethink the sizes of the images/ test the enviroment and resize the buttons
https://www.paintcodeapp.com/news/ultimate-guide-to-iphone-resolutions we can see that we have 320px in y for iPhone 5 (it's the same for iPhone SE).
The images are currently 90px. We have 4 images, so we need 360px for the images only.
In addition to that we have the navigation bar, which is typically 44, so we end up with needing 404px in y direction.
So we need to reduce pixels in the y direction with at least:
404 -320 = 84px.
And we have 4 images so we can reduce by 21px, but let's do 22px since it's a even number.
This gives us 90 - 22 = 68px.
Then since we are mainly working on @2 size, lets change that image to 136px and (also update the text in the image so we can easily confirm that the correct image is used at run time)